Sanitize types and coding style across repository (#1)
Sanitize types and coding style across repository (#1)

This PR sanitizes the existing code base to use a modern, consistent
style of C programming. It also changes all uses of fixed width
primitives (i.e., `int64_t`) to use the project specific typedefs
instead (i.e., `csm_i64`).

A home script has been added but it is not functional yet.

The next step for the repository will be to add a working test harness
that can be used to verify the correctness of the virtual machine,
and then to begin adding tests for edge cases.

I'll add a list of TODOs in a separate issue.