Adjust linux64 link order and chasm invocation (#4)
Adjust linux64 link order and chasm invocation (#4)

This PR cleans up the project Makefile and fixes some issues that were
preventing me from compiling on a fresh linux64 installation.

The linux64 linker is sensitive to the order in which libraries are
passed via the command line. This was preventing driver programs
that referenced the VM shared library from linking against it. To
fix it, we link against the VM last (in other words, list the driver
source file first).

We also adjust the `chasm` invocation to be a proper call to
`python3`, and make sure to create the `dump` directory before
compiling any tests.